pyproject.toml
[tool.poetry]
name = "questionator"
version = "0.0.0"
description = "Private bot for making QOTD questions"
authors = ["No767 <73260931+No767@users.noreply.github.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
orjson = "^3.8.1"
aiodns = "^3.0.0"
Brotli = "^1.0.9"
python-dotenv = "^0.21.0"
"discord.py" = "^2.0.1"
uvloop = "^0.17.0"
charset-normalizer = "^2.1.1"
tortoise-orm = {extras = ["accel"], version = "^0.19.2"}
asyncpg = "^0.27.0"
coredis = "^4.8.0"
pynacl = "~1.5.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
pyinstrument = "^4.3.0"

[tool.isort]
profile = 'black'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
